+++ src/Views/MessageBox.axaml.cs
@@ -1,0 +1,90 @@
+﻿using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+using Avalonia.Platform;
+using System;
+using System.IO;
+using System.Threading.Tasks;
+
+namespace NAND_Extractor.Views
+{
+    class MessageBox : Window
+    {
+        public enum MessageBoxButtons
+        {
+            Ok,
+            OkCancel,
+            YesNo,
+            YesNoCancel
+        }
+
+        public enum MessageBoxResult
+        {
+            Ok,
+            Cancel,
+            Yes,
+            No
+        }
+
+        public MessageBox()
+        {
+            this.InitializeComponent();
+#if DEBUG
+            this.AttachDevTools();
+#endif
+        }
+
+        private void InitializeComponent()
+        {
+            AvaloniaXamlLoader.Load(this);
+        }
+
+        public static Task<MessageBoxResult> Show(Window parent, string text, string title, MessageBoxButtons buttons, string icon)
+        {
+            var assets = AvaloniaLocator.Current.GetService<IAssetLoader>();
+            var assemblyName = typeof(MessageBox).Assembly.GetName().Name;
+            using var iconStream = assets.Open(new Uri($"avares://{assemblyName}/Assets/{icon}.ico"));
+            var msgbox = new MessageBox()
+            {
+                Title = title,
+                Icon = new WindowIcon(iconStream),
+            };
+
+            msgbox.FindControl<TextBlock>("Text").Text = text;
+            var buttonPanel = msgbox.FindControl<StackPanel>("Buttons");
+
+            var res = MessageBoxResult.Ok;
+
+            void AddButton(string caption, MessageBoxResult r, bool def = false)
+            {
+                var btn = new Button { Content = caption };
+                btn.Click += (_, __) => {
+                    res = r;
+                    msgbox.Close();
+                };
+                buttonPanel.Children.Add(btn);
+                if (def)
+                    res = r;
+            }
+
+            if (buttons == MessageBoxButtons.Ok || buttons == MessageBoxButtons.OkCancel)
+                AddButton("Ok", MessageBoxResult.Ok, true);
+            if (buttons == MessageBoxButtons.YesNo || buttons == MessageBoxButtons.YesNoCancel)
+            {
+                AddButton("Yes", MessageBoxResult.Yes);
+                AddButton("No", MessageBoxResult.No, true);
+            }
+
+            if (buttons == MessageBoxButtons.OkCancel || buttons == MessageBoxButtons.YesNoCancel)
+                AddButton("Cancel", MessageBoxResult.Cancel, true);
+
+
+            var tcs = new TaskCompletionSource<MessageBoxResult>();
+            msgbox.Closed += delegate { tcs.TrySetResult(res); };
+            if (parent != null)
+                msgbox.ShowDialog(parent);
+            else msgbox.Show();
+            return tcs.Task;
+        }
+    }
+}